+++ app/routers/datafilter.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING, Optional
 from fastapi import APIRouter, status
 from oteapi.models import FilterConfig
 from oteapi.plugins import create_strategy
+from oteapi.utils.config_updater import populate_config_from_session
 
 from app.models.datafilter import (
     IDPREFIX,
@@ -13,25 +14,22 @@ from app.models.datafilter import (
     GetFilterResponse,
     InitializeFilterResponse,
 )
-from app.models.error import HTTPNotFoundError, httpexception_404_item_id_does_not_exist
+from app.models.error import HTTPNotFoundError
 from app.redis_cache import TRedisPlugin
+from app.redis_cache._cache import _fetch_cache_value, _validate_cache_key
 from app.routers.session import _update_session, _update_session_list_item
 
 if TYPE_CHECKING:  # pragma: no cover
-    from typing import Any
-
     from oteapi.interfaces import IFilterStrategy
 
-ROUTER = APIRouter(prefix=f"/{IDPREFIX}")
-
-
-@ROUTER.post(
-    "/",
-    response_model=CreateFilterResponse,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
+ROUTER = APIRouter(
+    prefix=f"/{IDPREFIX}",
+    tags=["datafilter"],
+    responses={status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError}},
 )
+
+
+@ROUTER.post("/", response_model=CreateFilterResponse)
 async def create_filter(
     cache: TRedisPlugin,
     config: FilterConfig,
@@ -43,8 +41,7 @@ async def create_filter(
     await cache.set(new_filter.filter_id, config.model_dump_json())
 
     if session_id:
-        if not await cache.exists(session_id):
-            raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
+        await _validate_cache_key(cache, session_id, "session_id")
         await _update_session_list_item(
             session_id=session_id,
             list_key="filter_info",
@@ -55,45 +52,22 @@ async def create_filter(
     return new_filter
 
 
-@ROUTER.get(
-    "/{filter_id}",
-    response_model=GetFilterResponse,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
-)
+@ROUTER.get("/{filter_id}", response_model=GetFilterResponse)
 async def get_filter(
     cache: TRedisPlugin,
     filter_id: str,
     session_id: Optional[str] = None,
 ) -> GetFilterResponse:
     """Run and return data from a filter (data operation)"""
-    if not await cache.exists(filter_id):
-        raise httpexception_404_item_id_does_not_exist(filter_id, "filter_id")
-    if session_id and not await cache.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(filter_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {filter_id} to be a string or bytes, "
-            f"found it to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, filter_id, "filter_id")
     config = FilterConfig(**json.loads(cache_value))
 
-    strategy: "IFilterStrategy" = create_strategy("filter", config)
-
     if session_id:
-        cache_value = await cache.get(session_id)
-        if not isinstance(cache_value, (str, bytes)):
-            raise TypeError(
-                f"Expected cache value of {session_id} to be a string or bytes, "
-                f"found it to be of type {type(cache_value)!r}."
-            )
-    session_data: "Optional[dict[str, Any]]" = (
-        None if not session_id else json.loads(cache_value)
-    )
-    session_update = strategy.get(session=session_data)
+        session_data = await _fetch_cache_value(cache, session_id, "session_id")
+        populate_config_from_session(json.loads(session_data), config)
+
+    strategy: "IFilterStrategy" = create_strategy("filter", config)
+    session_update = strategy.get()
 
     if session_update and session_id:
         await _update_session(
@@ -103,45 +77,22 @@ async def get_filter(
     return GetFilterResponse(**session_update)
 
 
-@ROUTER.post(
-    "/{filter_id}/initialize",
-    response_model=InitializeFilterResponse,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
-)
+@ROUTER.post("/{filter_id}/initialize", response_model=InitializeFilterResponse)
 async def initialize_filter(
     cache: TRedisPlugin,
     filter_id: str,
     session_id: Optional[str] = None,
 ) -> InitializeFilterResponse:
     """Initialize and return data to update session."""
-    if not await cache.exists(filter_id):
-        raise httpexception_404_item_id_does_not_exist(filter_id, "filter_id")
-    if session_id and not await cache.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(filter_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {filter_id} to be a string or bytes, "
-            f"found it to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, filter_id, "filter_id")
     config = FilterConfig(**json.loads(cache_value))
 
-    strategy: "IFilterStrategy" = create_strategy("filter", config)
-
     if session_id:
-        cache_value = await cache.get(session_id)
-        if not isinstance(cache_value, (str, bytes)):
-            raise TypeError(
-                f"Expected cache value of {session_id} to be a string or bytes, "
-                f"found it to be of type {type(cache_value)!r}."
-            )
-    session_data: "Optional[dict[str, Any]]" = (
-        None if not session_id else json.loads(cache_value)
-    )
-    session_update = strategy.initialize(session=session_data)
+        session_data = await _fetch_cache_value(cache, session_id, "session_id")
+        populate_config_from_session(json.loads(session_data), config)
+
+    strategy: "IFilterStrategy" = create_strategy("filter", config)
+    session_update = strategy.initialize()
 
     if session_update and session_id:
         await _update_session(