+++ app/routers/dataresource.py
@@ -1,11 +1,12 @@
 """Data Resource."""
 
 import json
-from typing import TYPE_CHECKING, Optional
+from typing import Optional
 
 from fastapi import APIRouter, Request, status
 from oteapi.models import ResourceConfig
 from oteapi.plugins import create_strategy
+from oteapi.utils.config_updater import populate_config_from_session
 
 from app.models.dataresource import (
     IDPREFIX,
@@ -16,29 +17,27 @@ from app.models.dataresource import (
 from app.models.error import (
     HTTPNotFoundError,
     HTTPValidationError,
-    httpexception_404_item_id_does_not_exist,
     httpexception_422_resource_id_is_unprocessable,
 )
 from app.redis_cache import TRedisPlugin
+from app.redis_cache._cache import _fetch_cache_value, _validate_cache_key
 from app.routers.session import _update_session, _update_session_list_item
 
-if TYPE_CHECKING:  # pragma: no cover
-    from typing import Any
-
-ROUTER = APIRouter(prefix=f"/{IDPREFIX}")
-
-
-@ROUTER.post(
-    "/",
-    response_model=CreateResourceResponse,
+ROUTER = APIRouter(
+    prefix=f"/{IDPREFIX}",
+    tags=["dataresource"],
     responses={
         status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
+        status.HTTP_422_UNPROCESSABLE_ENTITY: {"model": HTTPValidationError},
     },
 )
+
+
+@ROUTER.post("/", response_model=CreateResourceResponse)
 async def create_dataresource(
     cache: TRedisPlugin,
-    config: ResourceConfig,
     request: Request,
+    config: ResourceConfig,
     session_id: Optional[str] = None,
 ) -> CreateResourceResponse:
     """### Register an external data resource.
@@ -63,8 +62,7 @@ async def create_dataresource(
     await cache.set(new_resource.resource_id, resource_config)
 
     if session_id:
-        if not await cache.exists(session_id):
-            raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
+        await _validate_cache_key(cache, session_id, "session_id")
         await _update_session_list_item(
             session_id=session_id,
             list_key="resource_info",
@@ -76,26 +74,14 @@ async def create_dataresource(
 
 
 @ROUTER.get(
-    "/{resource_id}/info",
-    response_model=ResourceConfig,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
+    "/{resource_id}/info", response_model=ResourceConfig, include_in_schema=False
 )
 async def info_dataresource(
     cache: TRedisPlugin,
     resource_id: str,
 ) -> ResourceConfig:
     """Get data resource info."""
-    if not await cache.exists(resource_id):
-        raise httpexception_404_item_id_does_not_exist(resource_id, "resource_id")
-
-    cache_value = await cache.get(resource_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {resource_id} to be a string or bytes, "
-            f"found it to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, resource_id, "resource_id")
     return ResourceConfig(**json.loads(cache_value))
 
 
@@ -103,10 +89,6 @@ async def info_dataresource(
     "/{resource_id}",
     response_model=GetResourceResponse,
     response_model_exclude_unset=True,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-        status.HTTP_422_UNPROCESSABLE_ENTITY: {"model": HTTPValidationError},
-    },
 )
 async def read_dataresource(
     cache: TRedisPlugin,
@@ -117,121 +99,56 @@ async def read_dataresource(
 
     Parse data information using the appropriate parser.
     """
-    if not await cache.exists(resource_id):
-        raise httpexception_404_item_id_does_not_exist(resource_id, "resource_id")
-    if session_id and not await cache.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(resource_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {resource_id} to be a string or bytes, "
-            f"found it to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, resource_id, "resource_id")
     config = ResourceConfig(**json.loads(cache_value))
 
-    if session_id:
-        cache_value = await cache.get(session_id)
-        if not isinstance(cache_value, (str, bytes)):
-            raise TypeError(
-                f"Expected cache value of {session_id} to be a string or bytes, "
-                f"found it to be of type {type(cache_value)!r}."
-            )
-    session_data: "Optional[dict[str, Any]]" = (
-        None if not session_id else json.loads(cache_value)
-    )
-
-    if config.downloadUrl and config.mediaType:
-        # Download strategy
-        session_update = create_strategy("download", config).get(session=session_data)
-        if session_update and session_id:
-            session_data = await _update_session(session_id, session_update, cache)
-
-        # Parse strategy
-        session_update = create_strategy("parse", config).get(session=session_data)
-        if session_update and session_id:
-            await _update_session(session_id, session_update, cache)
-
-    elif config.accessUrl and config.accessService:
-        # Resource strategy
-        session_update = create_strategy("resource", config).get(session=session_data)
-        if session_update and session_id:
-            await _update_session(session_id, session_update, cache)
-
-    else:
+    if (not config.resourceType) or (
+        not (
+            (config.downloadUrl and config.mediaType)
+            or (config.accessUrl and config.accessService)
+        )
+    ):
         raise httpexception_422_resource_id_is_unprocessable(resource_id)
+
+    if session_id:
+        session_data = await _fetch_cache_value(cache, session_id, "session_id")
+        populate_config_from_session(json.loads(session_data), config)
+
+    session_update = create_strategy("resource", config).get()
+    if session_update and session_id:
+        await _update_session(
+            session_id=session_id, updated_session=session_update, redis=cache
+        )
 
     return GetResourceResponse(**session_update)
 
 
-@ROUTER.post(
-    "/{resource_id}/initialize",
-    response_model=InitializeResourceResponse,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-        status.HTTP_422_UNPROCESSABLE_ENTITY: {"model": HTTPValidationError},
-    },
-)
+@ROUTER.post("/{resource_id}/initialize", response_model=InitializeResourceResponse)
 async def initialize_dataresource(
     cache: TRedisPlugin,
     resource_id: str,
     session_id: Optional[str] = None,
 ) -> InitializeResourceResponse:
     """Initialize data resource."""
-    if not await cache.exists(resource_id):
-        raise httpexception_404_item_id_does_not_exist(resource_id, "resource_id")
-    if session_id and not await cache.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(resource_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {resource_id} to be a string or bytes, "
-            f"found it to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, resource_id, "resource_id")
     config = ResourceConfig(**json.loads(cache_value))
 
-    if session_id:
-        cache_value = await cache.get(session_id)
-        if not isinstance(cache_value, (str, bytes)):
-            raise TypeError(
-                f"Expected cache value of {session_id} to be a string or bytes, "
-                f"found it to be of type {type(cache_value)!r}."
-            )
-    session_data: "Optional[dict[str, Any]]" = (
-        None if not session_id else json.loads(cache_value)
-    )
-
-    if config.downloadUrl and config.mediaType:
-        # Download strategy
-        session_update = create_strategy("download", config).initialize(
-            session=session_data
+    if (not config.resourceType) or (
+        not (
+            (config.downloadUrl and config.mediaType)
+            or (config.accessUrl and config.accessService)
         )
-        if session_update and session_id:
-            session_data = await _update_session(
-                session_id=session_id, updated_session=session_update, redis=cache
-            )
-
-        # Parse strategy
-        session_update = create_strategy("parse", config).initialize(
-            session=session_data
-        )
-        if session_update and session_id:
-            session_data = await _update_session(
-                session_id=session_id, updated_session=session_update, redis=cache
-            )
-
-    elif config.accessUrl and config.accessService:
-        # Resource strategy
-        session_update = create_strategy("resource", config).initialize(
-            session=session_data
-        )
-        if session_update and session_id:
-            session_data = await _update_session(
-                session_id=session_id, updated_session=session_update, redis=cache
-            )
-
-    else:
+    ):
         raise httpexception_422_resource_id_is_unprocessable(resource_id)
+
+    if session_id:
+        session_data = await _fetch_cache_value(cache, session_id, "session_id")
+        populate_config_from_session(json.loads(session_data), config)
+
+    session_update = create_strategy("resource", config).initialize()
+    if session_update and session_id:
+        await _update_session(
+            session_id=session_id, updated_session=session_update, redis=cache
+        )
 
     return InitializeResourceResponse(**session_update)