+++ app/routers/session.py
@@ -4,9 +4,9 @@ import json
 from typing import TYPE_CHECKING
 
 from fastapi import APIRouter, status
-from oteapi.models import SessionUpdate
+from oteapi.models import AttrDict
 
-from app.models.error import HTTPNotFoundError, httpexception_404_item_id_does_not_exist
+from app.models.error import HTTPNotFoundError
 from app.models.response import Session
 from app.models.session import (
     IDPREFIX,
@@ -16,19 +16,20 @@ from app.models.session import (
     ListSessionsResponse,
 )
 from app.redis_cache import TRedisPlugin
+from app.redis_cache._cache import _fetch_cache_value, _validate_cache_key
 
 if TYPE_CHECKING:  # pragma: no cover
     from typing import Any, Union
 
 
-ROUTER = APIRouter(prefix=f"/{IDPREFIX}")
-
-
-@ROUTER.post(
-    "/",
-    response_model=CreateSessionResponse,
+ROUTER = APIRouter(
+    prefix=f"/{IDPREFIX}",
+    tags=["session"],
     responses={status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError}},
 )
+
+
+@ROUTER.post("/", response_model=CreateSessionResponse)
 async def create_session(
     cache: TRedisPlugin,
     session: Session,
@@ -48,36 +49,30 @@ async def create_session(
 
     """
     new_session = CreateSessionResponse()
-
     await cache.set(new_session.session_id, session.model_dump_json())
     return new_session
 
 
-@ROUTER.get(
-    "/",
-    response_model=ListSessionsResponse,
-    responses={status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError}},
-)
+@ROUTER.get("/", response_model=ListSessionsResponse)
 async def list_sessions(
     cache: TRedisPlugin,
 ) -> ListSessionsResponse:
     """Get all session keys"""
-    keylist = []
+    keylist: list[str] = []
     for key in await cache.keys(pattern=f"{IDPREFIX}*"):
-        if not isinstance(key, bytes):
+        if not isinstance(key, (str, bytes)):
             raise TypeError(
                 "Found a key that is not stored as bytes (stored as type "
                 f"{type(key)!r})."
             )
-        keylist.append(key.decode(encoding="utf-8"))
+        if isinstance(key, bytes):
+            key = key.decode(encoding="utf-8")
+        keylist.append(key)
+
     return ListSessionsResponse(keys=keylist)
 
 
-@ROUTER.delete(
-    "/",
-    response_model=DeleteAllSessionsResponse,
-    responses={status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError}},
-)
+@ROUTER.delete("/", response_model=DeleteAllSessionsResponse)
 async def delete_all_sessions(
     cache: TRedisPlugin,
 ) -> DeleteAllSessionsResponse:
@@ -88,6 +83,12 @@ async def delete_all_sessions(
 
     """
     keylist = await cache.keys(pattern=f"{IDPREFIX}*")
+    # Check if the keylist is empty
+    if not keylist:
+        return DeleteAllSessionsResponse(
+            number_of_deleted_sessions=0, message="No session keys found to delete."
+        )
+
     return DeleteAllSessionsResponse(
         number_of_deleted_sessions=await cache.delete(*keylist)
     )
@@ -98,20 +99,13 @@ async def _get_session(
     redis: TRedisPlugin,
 ) -> Session:
     """Return the session contents given a `session_id`."""
-    if not await redis.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-    cache_value = await redis.get(session_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {session_id} to be a string or bytes, found it "
-            f"to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(redis, session_id, "session_id")
     return Session(**json.loads(cache_value))
 
 
 async def _update_session(
     session_id: str,
-    updated_session: "Union[SessionUpdate, dict[str, Any]]",
+    updated_session: "Union[AttrDict, dict[str, Any]]",
     redis: TRedisPlugin,
 ) -> Session:
     """Update an existing session (to be called internally)."""
@@ -149,74 +143,36 @@ async def _update_session_list_item(
     return session
 
 
-@ROUTER.put(
-    "/{session_id}",
-    response_model=Session,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
-)
+@ROUTER.put("/{session_id}", response_model=Session)
 async def update_session(
     cache: TRedisPlugin,
     session_id: str,
-    updated_session: SessionUpdate,
+    updated_session: AttrDict,
 ) -> Session:
     """Update session object."""
-    if not await cache.exists(session_id):
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(session_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {session_id} to be a string or bytes, found it "
-            f"to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, session_id, "session_id")
     session = Session(**json.loads(cache_value))
     session.update(updated_session)
     await cache.set(session_id, session.model_dump_json().encode("utf-8"))
     return session
 
 
-@ROUTER.get(
-    "/{session_id}",
-    response_model=Session,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
-)
+@ROUTER.get("/{session_id}", response_model=Session)
 async def get_session(
     cache: TRedisPlugin,
     session_id: str,
 ) -> Session:
     """Fetch the entire session object."""
-    cache_exists: int = await cache.exists(session_id)
-    if not cache_exists:
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
-    cache_value = await cache.get(session_id)
-    if not isinstance(cache_value, (str, bytes)):
-        raise TypeError(
-            f"Expected cache value of {session_id} to be a string or bytes, found it "
-            f"to be of type {type(cache_value)!r}."
-        )
+    cache_value = await _fetch_cache_value(cache, session_id, "session_id")
     return Session(**json.loads(cache_value))
 
 
-@ROUTER.delete(
-    "/{session_id}",
-    response_model=DeleteSessionResponse,
-    responses={
-        status.HTTP_404_NOT_FOUND: {"model": HTTPNotFoundError},
-    },
-)
+@ROUTER.delete("/{session_id}", response_model=DeleteSessionResponse)
 async def delete_session(
     cache: TRedisPlugin,
     session_id: str,
 ) -> DeleteSessionResponse:
     """Delete a session object."""
-    cache_exists: int = await cache.exists(session_id)
-    if not cache_exists:
-        raise httpexception_404_item_id_does_not_exist(session_id, "session_id")
-
+    await _validate_cache_key(cache, session_id, "session_id")
     await cache.delete(session_id)
     return DeleteSessionResponse(success=True)