+++ tests/routers/test_parser.py
@@ -1,4 +1,4 @@
-"""Test mapping."""
+""" Test parser """
 
 from typing import TYPE_CHECKING
 
@@ -10,39 +10,54 @@ if TYPE_CHECKING:
     from fastapi.testclient import TestClient
 
 
-def test_create_mapping(client: "TestClient") -> None:
-    """Test creating a mapping."""
+def test_create_parser(client: "TestClient") -> None:
+    """Test creating a parser"""
     response = client.post(
-        "/mapping/",
+        "/parser/",
         json={
-            "mappingType": "mapping/demo",
-            "prefixes": {},
-            "triples": [["a", "b", "c"]],
-            "configuration": {},
+            "parserType": "parser/demo",
+            "entity": "http://example.com/entity",
+            "configuration": {
+                "downloadUrl": "https://filesamples.com/sample2.json",
+                "mediaType": "application/json",
+            },
         },
         headers={"Content-Type": "application/json"},
         timeout=(3.0, 27.0),
     )
-    assert "mapping_id" in response.json()
+    assert "parser_id" in response.json()
     assert response.status_code == 200
 
 
-def test_get_mapping(client: "TestClient", test_data: dict[str, str]) -> None:
-    """Test getting a mapping."""
-    mapping_id = next(_ for _ in test_data if _.startswith("mapping-"))
+def test_info_parser(client: "TestClient", test_data: dict[str, str]) -> None:
+    """Test getting information about a parser"""
+    parser_id = next(_ for _ in test_data if _.startswith("parser-"))
     response = client.get(
-        f"/mapping/{mapping_id}",
+        f"/parser/{parser_id}/info",
+        headers={"Content-Type": "application/json"},
+        timeout=(3.0, 27.0),
+    )
+    assert response.status_code == 200
+    assert "parserType" in response.json()
+    assert "configuration" in response.json()
+
+
+def test_get_parser(client: "TestClient", test_data: dict[str, str]) -> None:
+    """Test getting and parsing data using a specified parser"""
+    parser_id = next(_ for _ in test_data if _.startswith("parser-"))
+    response = client.get(
+        f"/parser/{parser_id}",
         headers={"Content-Type": "application/json"},
         timeout=(3.0, 27.0),
     )
     assert response.status_code == 200
 
 
-def test_initialize_mapping(client: "TestClient", test_data: dict[str, str]) -> None:
-    """Test initializing a mapping."""
-    mapping_id = next(_ for _ in test_data if _.startswith("mapping-"))
+def test_initialize_parser(client: "TestClient", test_data: dict[str, str]) -> None:
+    """Test initializing a parser."""
+    parser_id = next(_ for _ in test_data if _.startswith("parser-"))
     response = client.post(
-        f"/mapping/{mapping_id}/initialize",
+        f"/parser/{parser_id}/initialize",
         headers={"Content-Type": "application/json"},
         timeout=(3.0, 27.0),
     )
@@ -59,23 +74,23 @@ def test_session_config_merge(
     """Test the current session is merged into the strategy configuration."""
     import json
 
-    from oteapi.models import MappingConfig
+    from oteapi.models import ParserConfig
     from oteapi.plugins import create_strategy
 
     original_create_strategy = create_strategy
 
-    mapping_id = next(_ for _ in test_data if _.startswith("mapping-"))
+    parser_id = next(_ for _ in test_data if _.startswith("parser-"))
     session_id = next(_ for _ in test_data if _.startswith("session-"))
 
-    expected_merged_config = json.loads(test_data[mapping_id])
+    expected_merged_config = json.loads(test_data[parser_id])
     expected_merged_config["configuration"].update(json.loads(test_data[session_id]))
 
     def create_strategy_middleware(
-        strategy_type: 'Literal["mapping"]', config: MappingConfig
+        strategy_type: 'Literal["parse"]', config: ParserConfig
     ):
         """Create a strategy middleware - do some testing."""
-        assert strategy_type == "mapping"
-        assert isinstance(config, MappingConfig)
+        assert strategy_type == "parse"
+        assert isinstance(config, ParserConfig)
 
         # THIS is where we test the session has been properly merged into the strategy
         # configuration !
@@ -86,19 +101,19 @@ def test_session_config_merge(
         return original_create_strategy(strategy_type, config)
 
     monkeypatch.setattr(
-        "app.routers.mapping.create_strategy", create_strategy_middleware
+        "app.routers.parser.create_strategy", create_strategy_middleware
     )
 
     if method == "initialize":
         response = client.post(
-            f"/mapping/{mapping_id}/initialize",
+            f"/parser/{parser_id}/initialize",
             params={"session_id": session_id},
             headers={"Content-Type": "application/json"},
             timeout=(3.0, 27.0),
         )
     else:  # method == "get"
         response = client.get(
-            f"/mapping/{mapping_id}",
+            f"/parser/{parser_id}",
             params={"session_id": session_id},
             headers={"Content-Type": "application/json"},
             timeout=(3.0, 27.0),